pyproject.toml
[project]
name = "16-bus-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for the 16 Bus SA taxi rideshare platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14,<2.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic-settings>=2.3,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "redis>=5.0,<6.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
]

[tool.alembic]
script_location = "alembic"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
